+++ hutool-core/src/test/java/cn/hutool/core/io/FileCopierTest.java
@@ -21,6 +21,16 @@ public class FileCopierTest {
 	
 	@Test
 	@Ignore
+	public void dirCopyTest2() {
+		//测试带.的文件夹复制
+		FileCopier copier = FileCopier.create("D:\\workspace\\java\\.metadata", "D:\\workspace\\java\\temp");
+		copier.copy();
+		
+		FileUtil.copy("D:\\workspace\\java\\looly\\hutool\\.git", "D:\\workspace\\java\\temp", true);
+	}
+	
+	@Test
+	@Ignore
 	public void copyFileToDirTest() {
 		FileCopier copier = FileCopier.create("d:/GReen_Soft/XshellXftpPortable.zip", "c:/hp/");
 		copier.copy();