+++ hutool-cron/src/main/java/cn/hutool/cron/TaskExecutorManager.java
@@ -3,7 +3,6 @@ package cn.hutool.cron;
 import java.util.ArrayList;
 import java.util.List;
 
-import cn.hutool.core.thread.ThreadUtil;
 import cn.hutool.cron.task.Task;
 
 /**
@@ -18,13 +17,14 @@ public class TaskExecutorManager {
 	protected Scheduler scheduler;
 	/** 执行器列表 */
 	private List<TaskExecutor> executors = new ArrayList<>();
-	
+
 	public TaskExecutorManager(Scheduler scheduler) {
 		this.scheduler = scheduler;
 	}
-	
+
 	/**
 	 * 启动 TaskExecutor
+	 * 
 	 * @param task {@link Task}
 	 * @return {@link TaskExecutor}
 	 */
@@ -33,13 +33,15 @@ public class TaskExecutorManager {
 		synchronized (this.executors) {
 			this.executors.add(executor);
 		}
-		executor.setDaemon(this.scheduler.daemon);
+		// 子线程是否为deamon线程取决于父线程，因此此处无需显示调用
+		// executor.setDaemon(this.scheduler.daemon);
 		executor.start();
 		return executor;
 	}
-	
+
 	/**
 	 * 执行器执行完毕调用此方法，将执行器从执行器列表移除
+	 * 
 	 * @param executor 执行器 {@link TaskExecutor}
 	 * @return this
 	 */
@@ -52,14 +54,17 @@ public class TaskExecutorManager {
 
 	/**
 	 * 停止所有TaskExecutor
+	 * 
 	 * @return this
+	 * @deprecated 作业执行器只是执行给定的定时任务线程，无法强制关闭，可通过deamon线程方式关闭之
 	 */
+	@Deprecated
 	public TaskExecutorManager destroy() {
-		synchronized (this.executors) {
-			for (TaskExecutor taskExecutor : executors) {
-				ThreadUtil.interupt(taskExecutor, true);
-			}
-		}
+		// synchronized (this.executors) {
+		// for (TaskExecutor taskExecutor : executors) {
+		// ThreadUtil.interupt(taskExecutor, false);
+		// }
+		// }
 		this.executors.clear();
 		return this;
 	}