+++ site/commit.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-git add .
-git commit -am "$1"
-./push.sh