+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 
 -------------------------------------------------------------------------------------------------------------
 
+## 4.1.3
+
+### 新特性
+* 【all】             优化db的DsFactory、log的LogFactory、extra的TemplateUtil逻辑，减少异常栈嵌套
+* 【core】          Validator增加isMac、validateMac方法（感谢@【上海】阳仔）
+
+### Bug修复
+* 【core】          修复ArrayUtil.join前后fix失效问题（@【河北】理想主义）
+* 【core】          修复DateRange最后一个元素逻辑问题（issue#ILE38@Gitee）
+* 【cron】          修复调用CronUtil.stop()方法无法正常结束作业进程的问题（issue#ILFCZ@Gitee）
+* 【db】             修复page方法在Oracle中丢失参数问题（issue#ILGXP@Gitee）
+* 【extra】          修复QrCodeUtil.decode对复杂二维码解码失败问题（感谢@【成都】小朋友）
+
+-------------------------------------------------------------------------------------------------------------
+
 ## 4.1.2
 
 ### 新特性